3.4/n example overhaul: dbt_python_assets to recommended project layout + rename (#9122)
* example overhaul: dbt_python_assets to recommended project layout

* remove old example

* isort

* no more need [tests] in setup.py

* update AVAILABLE_EXAMPLES
