+++ examples/assets_dbt_python/README.md
@@ -1,0 +1,25 @@
+# Using Python with dbt project Example
+
+This example demonstrates using Python alongside a medium-sized dbt project.
+
+## Getting started
+
+Bootstrap your own Dagster project with this example:
+
+```bash
+dagster project from-example --name my-dagster-project --example assets_dbt_python
+```
+
+To install this example and its Python dependencies, run:
+
+```bash
+pip install -e .
+```
+
+Once you've done this, you can run:
+
+```
+dagit
+```
+
+to view this example in Dagster's UI, Dagit.
