+++ examples/assets_dbt_python/assets_dbt_python/__init__.py
@@ -1,0 +1,1 @@
+from .repository import assets_dbt_python
