+++ examples/assets_dbt_python/assets_dbt_python/assets/raw_data/__init__.py
@@ -1,9 +1,8 @@
 import numpy as np
 import pandas as pd
+from assets_dbt_python.utils import random_data
 
 from dagster import asset
-
-from ..utils import random_data
 
 
 @asset(compute_kind="random")
