+++ examples/assets_dbt_python/assets_dbt_python/repository.py
@@ -1,9 +1,8 @@
 import os
 
-import dbt_python_assets.forecasting as forecasting
-import dbt_python_assets.raw_data as raw_data
+from assets_dbt_python.assets import forecasting, raw_data
+from assets_dbt_python.resources import duckdb_io_manager
 from dagster_dbt import dbt_cli_resource, load_assets_from_dbt_project
-from dbt_python_assets.resources import duckdb_io_manager
 
 from dagster import (
     ScheduleDefinition,
@@ -47,7 +46,7 @@ forecast_job = define_asset_job("refresh_forecast_model_job", selection="*order_
 
 
 @repository
-def example_repo():
+def assets_dbt_python():
     return with_resources(
         dbt_assets + raw_data_assets + forecasting_assets,
         resource_defs={
