+++ examples/assets_dbt_python/assets_dbt_python/resources/duckdb_io_manager.py
