+++ examples/assets_dbt_python/assets_dbt_python_tests/__init__.py
@@ -1,0 +1,1 @@
+
