+++ examples/assets_dbt_python/assets_dbt_python_tests/test_assets.py
@@ -1,0 +1,1 @@
+
