+++ examples/assets_dbt_python/dbt_project/.gitignore
