+++ examples/assets_dbt_python/dbt_project/config/profiles.yml
