+++ examples/assets_dbt_python/dbt_project/models/analytics/company_stats.sql
