+++ examples/assets_dbt_python/dbt_project/models/analytics/daily_order_summary.sql
