+++ examples/assets_dbt_python/dbt_project/models/analytics/order_stats.sql
