+++ examples/assets_dbt_python/dbt_project/models/analytics/orders_augmented.sql
