+++ examples/assets_dbt_python/dbt_project/models/analytics/sku_stats.sql
