+++ examples/assets_dbt_python/dbt_project/models/cleaned/orders_cleaned.sql
