+++ examples/assets_dbt_python/dbt_project/models/cleaned/schema.yml
