+++ examples/assets_dbt_python/dbt_project/models/cleaned/users_cleaned.sql
