+++ examples/assets_dbt_python/dbt_project/models/forecasting/all_orders.sql
