+++ examples/assets_dbt_python/dbt_project/models/marketing/top_users.sql
