+++ examples/assets_dbt_python/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = assets_dbt_python
