+++ examples/assets_dbt_python/tox.ini
@@ -2,8 +2,9 @@
 envlist = py{38, 37},pylint,mypy
 
 [testenv]
+usedevelop = true
 setenv =
-  VIRTUALENV_PIP=21.3.1
+  VIRTUALENV_PIP=22.1.2
 passenv = CI_* COVERALLS_REPO_TOKEN BUILDKITE
 deps =
   -e ../../python_modules/dagster[mypy,test]
@@ -23,4 +24,4 @@ commands =
 
 [testenv:pylint]
 commands =
-  pylint -j0 --rcfile=../../pyproject.toml {posargs} dbt_python_assets dbt_python_assets
+  pylint -j0 --rcfile=../../pyproject.toml {posargs} assets_dbt_python assets_dbt_python_tests
