+++ examples/assets_dbt_python/workspace.yaml
@@ -1,0 +1,2 @@
+load_from:
+  - python_package: assets_dbt_python
