--- examples/dbt_python_assets/README.md
@@ -1,19 +1,0 @@
-# Description
-
-This repo serves as an example of using Python alongside a medium-sized dbt project.
-
-# Setup
-
-To install this example and its Python dependencies, run:
-
-```
-$ pip install -e .
-```
-
-Once you've done this, you can run:
-
-```
-$ dagit -m dbt_python_assets
-```
-
-to view this example in Dagster's UI, Dagit.
