--- examples/dbt_python_assets/dbt_python_assets/__init__.py
@@ -1,1 +1,0 @@
-from .repo import example_repo
